Consolidate metadataStoreUtils
